Move changes from lsm_hooks.h to security.c, and use xmacro to copy the init the static calls
